# Optimize `base64.encodeMime`