* 5x faster for common scenarios, 13x faster if `lineLen` <= encoded string's length or `newLine` is empty.
* Changed `lineLen`'s type to `Positive` to disallow `0`.
```
name ............................... min time      avg time    std dv   runs
* `lineLen` >= encoded string's length:
new ................................ 0.001 ms      0.001 ms    ±0.000  x1000
old ................................ 0.013 ms      0.013 ms    ±0.000  x1000
* empty `newLine`:
new ................................ 0.001 ms      0.001 ms    ±0.000  x1000
old ................................ 0.014 ms      0.014 ms    ±0.001  x1000
* none of the above:
new ................................ 0.003 ms      0.003 ms    ±0.001  x1000
old ................................ 0.014 ms      0.015 ms    ±0.001  x1000
```