@AmjadHD 2022-09-23
I wanted to add:
```nim
doAssertRaises(RangeDefect): encodeMime("abc", lineLen = 0)
```
but compile time checks prevent this too early.