@AmjadHD 2022-09-23
Should `encodeMime` get a `safe` parameter ?
Should it accept an `openArray` like `encode` ?