@Araq 2022-09-23
Not my point, please ensure that `tests/base64/tbase64.nim` is sufficient or else improve it.