@AmjadHD 2022-09-23
it's sufficient.
Adding a `safe` parameter or supporting `openArray` makes it insufficient.